+++ example/lib/main.dart
@@ -1,6 +1,3 @@
-import 'dart:developer';
-import 'dart:math';
-
 import 'package:example/eloquents/car.dart';
 import 'package:example/eloquents/class.dart';
 import 'package:example/eloquents/student.dart';