+++ test/wazeloquent_test.dart
@@ -1,7 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:wazeloquent/wazeloquent.dart';
+// import 'package:wazeloquent/wazeloquent.dart';
 
 void main() {
-  test('adds one to input values', () {});
+  test('adds one to input values', () {
+    expect(true, true);
+  });
 }